requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
redis==5.0.1
python-dotenv==1.0.0
opencv-python==4.9.0.80
numpy==1.24.3
deepface==0.0.79
dlib==19.24.2 